+++ src/components/accordion-header/accordion-header.tsx
@@ -1,0 +1,16 @@
+import { Component, h, Host } from '@stencil/core';
+
+@Component({
+    tag: 'wcs-accordion-header',
+    // styleUrl: 'accordion-header.scss',
+    shadow: true
+})
+export class AccordionHeader {
+    render() {
+        return (
+            <Host slot="header">
+                <slot />
+            </Host>
+        );
+    }
+}